wren-agentic-system/pyproject.toml
[tool.poetry]
name = "wren-agentic-system"
version = "0.8.1"
description = ""
authors = ["Jimmy Yeh <jimmy.yeh@cannerdata.com>", "Pao Sheng Wang <david.wang@cannerdata.com>", "Aster Sun <aster.sun@cannerdata.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12.*, <4.0"
fastapi = "==0.109.2"
uvicorn = {extras = ["standard"], version = "==0.30.1"}
python-dotenv = "==1.0.1"
haystack-ai = "==2.3.1"
openai = "==1.30.1"
qdrant-haystack = "==3.5.0"
backoff = "==2.2.1"
tqdm = "==4.66.4"
numpy = "==1.26.4"
sqlparse = "==0.5.0"
sqlglot = "==22.5.0"
orjson = "==3.10.3"
sf-hamilton = {version = "==1.69.0", extras = ["visualization"]}
aiohttp = {extras = ["speedups"], version = "==3.9.5"}
ollama-haystack = "==0.0.6"
langfuse = "==2.35.0"
ollama = "==0.2.1"
toml = "==0.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "==8.2.0"
pre-commit = "==3.7.1"
pytest-cov = "==4.1.0"
gdown = "==5.2.0"
httpx = "==0.27.0"
streamlit = "==1.34.0"
watchdog = "==4.0.0"
ragas = "==0.1.1"
ragas-haystack = "==0.1.3"
psycopg2-binary = "==2.9.9"
setuptools = "==70.0.0"
locust = "==2.28.0"
pytest-asyncio = "==0.23.8"

[tool.poetry.group.eval.dependencies]
tomlkit = "==0.13.0"
deepeval = "==0.21.68"
streamlit-tags = "==1.2.8"
gitpython = "==3.1.43"
pandas = "==2.2.2"

[tool.poetry.group.demo.dependencies]
requests = "==2.32.2"


[tool.poetry.group.test.dependencies]
aioresponses = "==0.7.6"
pytest-mock = "==3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
